requirements.txt
asgiref==3.3.4
cffi==1.14.5
cryptography==3.4.7
dj-database-url==0.5.0
Django==3.2
django-cors-headers==3.7.0
django-heroku==0.3.1
djangorestframework==3.12.4
psycopg2==2.8.6
pycparser==2.20
PyJWT==1.7.1
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.2.0
